+++ Anarchy/REST/Guild/Settings/DiscordChannelSettings.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Discord
+{
+    public class DiscordChannelSettings : Controllable
+    {
+        [JsonProperty("muted")]
+        public bool Muted { get; private set; }
+    }
+}