+++ Anarchy/Voice/DiscordVoiceStream.cs
@@ -100,20 +100,7 @@ namespace Discord.Voice
 
         public int CopyFrom(string filePath, int offset = 0)
         {
-            var process = Process.Start(new ProcessStartInfo
-            {
-                FileName = "ffmpeg.exe",
-                Arguments = $"-hide_banner -loglevel panic -i \"{filePath}\" -ac 2 -f s16le -ar 48000 pipe:1",
-                UseShellExecute = false,
-                RedirectStandardOutput = true,
-            });
-
-            using (MemoryStream memStream = new MemoryStream())
-            {
-                process.StandardOutput.BaseStream.CopyTo(memStream);
-
-                return CopyFrom(memStream.ToArray(), offset);
-            }
+            return CopyFrom(DiscordVoiceUtils.ReadFromFile(filePath), offset);
         }
     }
 }